fix: add a git cli version
Workaround for: https://github.com/rust-lang/git2-rs/issues/1057
