+++ modules/importer/Cargo.toml
@@ -16,7 +16,7 @@ anyhow = { workspace = true }
 csaf = { workspace = true }
 csaf-walker = { workspace = true, features = ["crypto-openssl", "csaf"] }
 cve = { workspace = true }
-git2 = { workspace = true }
+git2 = { workspace = true, optional = true }
 hex = { workspace = true }
 humantime-serde = { workspace = true }
 log = { workspace = true }
@@ -42,6 +42,7 @@ utoipa = { workspace = true, features = ["actix_extras", "time", "url"] }
 uuid = { workspace = true, features = ["v4"] }
 walkdir = { workspace = true }
 walker-common = { workspace = true }
+xshell = { workspace = true }
 
 [dev-dependencies]
 test-log = { workspace = true, features = ["log", "trace"] }
